more registry stuff done
